# tests/api/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "CompactStatsResponse": {
        "description": "Zonal statistics for one pourpoint/dataset query: the echoed query, links,\nand the compact zone/variable/value body, in one schema shared by every\ndataset.",
        "properties": {
          "dataset": {
            "examples": [
              "snodas"
            ],
            "title": "Dataset",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "title": "Links",
            "type": "array"
          },
          "pourpoint": {
            "examples": [
              "12354500:MT:USGS"
            ],
            "pattern": "^[a-zA-Z0-9\\-]+:[a-zA-Z]{2}:[a-zA-Z]+$",
            "title": "Pourpoint",
            "type": "string"
          },
          "query": {
            "discriminator": {
              "mapping": {
                "DateRange": "#/components/schemas/DateRangeQuery",
                "DayOfYear": "#/components/schemas/DOYQuery"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/DateRangeQuery"
              },
              {
                "$ref": "#/components/schemas/DOYQuery"
              }
            ],
            "title": "Query"
          },
          "results": {
            "additionalProperties": {
              "items": {
                "items": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              "type": "array"
            },
            "examples": [
              {
                "2008-12-14": [
                  [
                    42.7
                  ],
                  [
                    51.3
                  ]
                ]
              }
            ],
            "propertyNames": {
              "format": "date"
            },
            "title": "Results",
            "type": "object"
          },
          "variables": {
            "examples": [
              [
                "mean_swe_mm"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Variables",
            "type": "array"
          },
          "zone_layers": {
            "examples": [
              [
                "terrain.elevation"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Zone Layers",
            "type": "array"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/CompactZone"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "zone_layers",
          "variables",
          "zones",
          "results",
          "pourpoint",
          "dataset",
          "query"
        ],
        "title": "CompactStatsResponse",
        "type": "object"
      },
...
      "DateRangeQuery": {
        "properties": {
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "2008-12-14"
            ],
            "title": "End Date"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "2008-12-01"
            ],
            "title": "Start Date"
          },
          "type": {
            "const": "DateRange",
            "default": "DateRange",
            "title": "Type",
            "type": "string"
          }
        },
        "title": "DateRangeQuery",
        "type": "object"
      },
...
      "FeatureCollection_PourpointProperties_": {
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/Feature_PourpointProperties_"
            },
            "title": "Items",
            "type": "array"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "title": "Links",
            "type": "array"
          },
          "numberMatched": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numbermatched"
          },
          "numberReturned": {
            "readOnly": true,
            "title": "Numberreturned",
            "type": "integer"
          },
          "type": {
            "const": "FeatureCollection",
            "default": "FeatureCollection",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "features",
          "numberReturned"
        ],
        "title": "FeatureCollection[PourpointProperties]",
        "type": "object"
      },
      "Feature_PourpointDetailProperties_": {
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "maxItems": 4,
                "minItems": 4,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "maxItems": 6,
                "minItems": 6,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "geometry": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "GeometryCollection": "#/components/schemas/GeometryCollection",
                    "LineString": "#/components/schemas/LineString",
                    "MultiLineString": "#/components/schemas/MultiLineString",
                    "MultiPoint": "#/components/schemas/MultiPoint",
                    "MultiPolygon": "#/components/schemas/MultiPolygon",
                    "Point": "#/components/schemas/Point",
                    "Polygon": "#/components/schemas/Polygon"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Point"
                  },
                  {
                    "$ref": "#/components/schemas/MultiPoint"
                  },
                  {
                    "$ref": "#/components/schemas/LineString"
                  },
                  {
                    "$ref": "#/components/schemas/MultiLineString"
                  },
                  {
                    "$ref": "#/components/schemas/Polygon"
                  },
                  {
                    "$ref": "#/components/schemas/MultiPolygon"
                  },
                  {
                    "$ref": "#/components/schemas/GeometryCollection"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "title": "Links",
            "type": "array"
          },
          "properties": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PourpointDetailProperties"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "const": "Feature",
            "default": "Feature",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "geometry",
          "properties"
        ],
        "title": "Feature[PourpointDetailProperties]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Link": {
        "additionalProperties": true,
        "description": "An OGC-style link. Null fields are omitted on JSON serialization.",
        "properties": {
          "body": {
            "title": "Body"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "href": {
            "$ref": "#/components/schemas/Url"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "rel": {
            "title": "Rel",
            "type": "string"
          },
          "templated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templated"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "required": [
          "href",
          "rel"
        ],
        "title": "Link",
        "type": "object"
      },
...
      "StatsFormat": {
        "description": "The requested output format, as one of the supported representation keys. Takes precedence over the `Accept` header; an unsupported value returns a `400` problem. Supported keys: `json`, `csv`.",
        "enum": [
          "json",
          "csv"
        ],
        "title": "StatsFormat",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Pourpoint metadata and snow zonal statistics.",
    "title": "PSU CSAR snowtool API",
    "version": "0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/datasets/{dataset}/stats/{triplet}/date-range": {
      "get": {
        "operationId": "stats_date_range_datasets__dataset__stats__triplet__date_range_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "triplet",
            "required": true,
            "schema": {
              "examples": [
                "12354500:MT:USGS"
              ],
              "pattern": "^[a-zA-Z0-9\\-]+:[a-zA-Z]{2}:[a-zA-Z]+$",
              "title": "Triplet",
              "type": "string"
            }
          },
          {
            "description": "Stratify by a zone layer (repeatable): LAYER or LAYER:PARAM=VALUE. See GET /datasets/{dataset} for the valid layer keys, override params, and variables. Default: whole basin.",
            "in": "query",
            "name": "zone",
            "required": false,
            "schema": {
              "description": "Stratify by a zone layer (repeatable): LAYER or LAYER:PARAM=VALUE. See GET /datasets/{dataset} for the valid layer keys, override params, and variables. Default: whole basin.",
              "items": {
                "type": "string"
              },
              "title": "Zone",
              "type": "array"
            }
          },
          {
            "description": "Variable to report (repeatable; default: all). Use a variable key from GET /datasets/{dataset}.",
            "in": "query",
            "name": "variable",
            "required": false,
            "schema": {
              "description": "Variable to report (repeatable; default: all). Use a variable key from GET /datasets/{dataset}.",
              "items": {
                "type": "string"
              },
              "title": "Variable",
              "type": "array"
            }
          },
          {
            "description": "Permit a basin only partially covered by the dataset grid (default false: a partially-covered basin is a 409). A wholly off-grid basin always 409s.",
            "in": "query",
            "name": "allow_partial",
            "required": false,
            "schema": {
              "default": false,
              "description": "Permit a basin only partially covered by the dataset grid (default false: a partially-covered basin is a 409). A wholly off-grid basin always 409s.",
              "title": "Allow Partial",
              "type": "boolean"
            }
          },
          {
            "description": "Include crossed zones that no AOI pixel falls in (0 area, all values null). By default these empty combinations are dropped. No effect on a whole-basin query.",
            "in": "query",
            "name": "include_empty_zones",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include crossed zones that no AOI pixel falls in (0 area, all values null). By default these empty combinations are dropped. No effect on a whole-basin query.",
              "title": "Include Empty Zones",
              "type": "boolean"
            }
          },
          {
            "description": "The requested output format, as one of the supported representation keys. Takes precedence over the `Accept` header; an unsupported value returns a `400` problem. Supported keys: `json`, `csv`.",
            "in": "query",
            "name": "f",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsFormat"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The requested output format, as one of the supported representation keys. Takes precedence over the `Accept` header; an unsupported value returns a `400` problem. Supported keys: `json`, `csv`.",
              "title": "F"
            }
          },
          {
            "description": "An RFC 3339 instant, or a `start/end` interval where either side may be open using `..` (or empty). An instant matches a single timestamp; an interval matches a range. Malformed values return a `400` problem.",
            "in": "query",
            "name": "datetime",
            "required": false,
            "schema": {
              "description": "An RFC 3339 instant, or a `start/end` interval where either side may be open using `..` (or empty). An instant matches a single timestamp; an interval matches a range. Malformed values return a `400` problem.",
              "example": "2018-01-01/2018-06-30",
              "examples": [
                "2018-01-01/2018-06-30",
                "2018-04-27",
                "2018-01-01/..",
                "../2018-06-30"
              ],
              "title": "Datetime",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactStatsResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stats Date Range",
        "tags": [
          "stats"
        ]
      }
    },
    "/datasets/{dataset}/stats/{triplet}/doy": {
      "get": {
        "operationId": "stats_doy_datasets__dataset__stats__triplet__doy_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "triplet",
            "required": true,
            "schema": {
              "examples": [
                "12354500:MT:USGS"
              ],
              "pattern": "^[a-zA-Z0-9\\-]+:[a-zA-Z]{2}:[a-zA-Z]+$",
              "title": "Triplet",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "examples": [
                4
              ],
              "maximum": 12,
              "minimum": 1,
              "title": "Month",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": true,
            "schema": {
              "examples": [
                1
              ],
              "maximum": 31,
              "minimum": 1,
              "title": "Day",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start_year",
            "required": true,
            "schema": {
              "examples": [
                2008
              ],
              "maximum": 9999,
              "minimum": 1,
              "title": "Start Year",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "end_year",
            "required": true,
            "schema": {
              "examples": [
                2020
              ],
              "maximum": 9999,
              "minimum": 1,
              "title": "End Year",
              "type": "integer"
            }
          },
          {
            "description": "Stratify by a zone layer (repeatable): LAYER or LAYER:PARAM=VALUE. See GET /datasets/{dataset} for the valid layer keys, override params, and variables. Default: whole basin.",
            "in": "query",
            "name": "zone",
            "required": false,
            "schema": {
              "description": "Stratify by a zone layer (repeatable): LAYER or LAYER:PARAM=VALUE. See GET /datasets/{dataset} for the valid layer keys, override params, and variables. Default: whole basin.",
              "items": {
                "type": "string"
              },
              "title": "Zone",
              "type": "array"
            }
          },
          {
            "description": "Variable to report (repeatable; default: all). Use a variable key from GET /datasets/{dataset}.",
            "in": "query",
            "name": "variable",
            "required": false,
            "schema": {
              "description": "Variable to report (repeatable; default: all). Use a variable key from GET /datasets/{dataset}.",
              "items": {
                "type": "string"
              },
              "title": "Variable",
              "type": "array"
            }
          },
          {
            "description": "Permit a basin only partially covered by the dataset grid (default false: a partially-covered basin is a 409). A wholly off-grid basin always 409s.",
            "in": "query",
            "name": "allow_partial",
            "required": false,
            "schema": {
              "default": false,
              "description": "Permit a basin only partially covered by the dataset grid (default false: a partially-covered basin is a 409). A wholly off-grid basin always 409s.",
              "title": "Allow Partial",
              "type": "boolean"
            }
          },
          {
            "description": "Include crossed zones that no AOI pixel falls in (0 area, all values null). By default these empty combinations are dropped. No effect on a whole-basin query.",
            "in": "query",
            "name": "include_empty_zones",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include crossed zones that no AOI pixel falls in (0 area, all values null). By default these empty combinations are dropped. No effect on a whole-basin query.",
              "title": "Include Empty Zones",
              "type": "boolean"
            }
          },
          {
            "description": "The requested output format, as one of the supported representation keys. Takes precedence over the `Accept` header; an unsupported value returns a `400` problem. Supported keys: `json`, `csv`.",
            "in": "query",
            "name": "f",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsFormat"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The requested output format, as one of the supported representation keys. Takes precedence over the `Accept` header; an unsupported value returns a `400` problem. Supported keys: `json`, `csv`.",
              "title": "F"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactStatsResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stats Doy",
        "tags": [
          "stats"
        ]
      }
    },
    "/pourpoints": {
      "get": {
        "operationId": "list_pourpoints_pourpoints_get",
        "parameters": [
          {
            "in": "query",
            "name": "geometry",
            "required": false,
            "schema": {
              "default": "point",
              "enum": [
                "point",
                "basin"
              ],
              "title": "Geometry",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "A bounding box, as comma-separated numbers `minx,miny,maxx,maxy` (WGS 84 longitude/latitude), or `minx,miny,minz,maxx,maxy,maxz` for a 3D box. The x axis may wrap the antimeridian (`minx` may exceed `maxx`); the y (and z) axes must be ordered `min <= max`. Malformed values return a `400` problem.",
            "examples": {
              "-180,-90,180,90": {
                "value": "-180,-90,180,90"
              },
              "7,50,0,8,51,1000": {
                "value": "7,50,0,8,51,1000"
              },
              "7.0,50.0,8.0,51.0": {
                "value": "7.0,50.0,8.0,51.0"
              }
            },
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A bounding box, as comma-separated numbers `minx,miny,maxx,maxy` (WGS 84 longitude/latitude), or `minx,miny,minz,maxx,maxy,maxz` for a 3D box. The x axis may wrap the antimeridian (`minx` may exceed `maxx`); the y (and z) axes must be ordered `min <= max`. Malformed values return a `400` problem.",
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureCollection_PourpointProperties_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pourpoints",
        "tags": [
          "pourpoints"
        ]
      }
    },
    "/pourpoints/{triplet}": {
      "get": {
        "operationId": "get_pourpoint_pourpoints__triplet__get",
        "parameters": [
          {
            "in": "path",
            "name": "triplet",
            "required": true,
            "schema": {
              "examples": [
                "12354500:MT:USGS"
              ],
              "pattern": "^[a-zA-Z0-9\\-]+:[a-zA-Z]{2}:[a-zA-Z]+$",
              "title": "Triplet",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature_PourpointDetailProperties_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pourpoint",
        "tags": [
          "pourpoints"
        ]
      }
    },
...
  },
...
}